% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_initialize.R
\name{easyRF.initialize}
\alias{easyRF.initialize}
\title{#########################################################################################
Initialize easyRF}
\usage{
easyRF.initialize()
}
\description{
Patrice Vallin, easyRF, Apr 2020.
########################################################################################
}
