% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_searchGrid.R
\name{generate_searchGrid}
\alias{generate_searchGrid}
\title{Generate searchGrid for RF}
\usage{
generate_searchGrid(ncol_DM, seed, rf_pars)
}
\arguments{
\item{ncol_DM}{a integer value}

\item{seed}{a integer value}

\item{rf_pars}{a list containing all rf pars}
}
\description{
Patrice Vallin, easyrf, Apr 2020
}
