% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_analysis.R
\name{rf_analysis}
\alias{rf_analysis}
\title{Analyse the dataset with randomForest}
\usage{
rf_analysis(dataFile)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Patrice Vallin, easyRF, Apr 2020
.
SSE and findElbow function are adapted from FlowSOM R package
Sofie Van Gassen et al., Cytometry Part A, 2015
https://onlinelibrary.wiley.com/doi/full/10.1002/cyto.a.22625
http://bioconductor.org/packages/release/bioc/html/FlowSOM.html
}
