% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_rf helper2.R
\name{rf_cv}
\alias{rf_cv}
\title{train cv rf model}
\usage{
rf_cv(data_DM, Y_data, seed, metric, control, searchGrid)
}
\arguments{
\item{data_DM}{a data matrix of numeric value}

\item{Y_data}{an array of numeric value}

\item{metric}{a character string refering to the name of the metric being autotuned}

\item{control}{a character string}

\item{searchGrid}{a matrix}
}
\description{
Patrice Vallin, easyrf, Apr 2020
}
