% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_helper.R
\name{rf_hypertuning}
\alias{rf_hypertuning}
\title{Find optimal xgboost parameters through hypertuning}
\usage{
rf_hypertuning(data_DM, searchGrid)
}
\arguments{
\item{data_DM}{an xgboost pointer}

\item{searchGrid}{a matrix}
}
\description{
Patrice Vallin, easyXgboost, Apr 2020
}
