% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_plot.R
\name{rf_plot}
\alias{rf_plot}
\title{Explain rf models}
\usage{
rf_plot(dataFile)
}
\arguments{
\item{dataFile}{a character string specifying the location of the project dataset}
}
\description{
Patrice Vallin, easyrf, Apr 2020
}
\details{
Plots were adapted from  https://www.kaggle.com/pradeeptripathi/prediction-of-titanic-survival-using-random-forest
}
